fix(ui): align overview session labels
