+++ ui/src/ui/views/overview.render.test.ts
@@ -134,4 +134,50 @@ describe("overview view rendering", () => {
       "openclaw devices list",
     ]);
   });
+
+  it("renders recent session names through the shared display resolver", async () => {
+    const container = document.createElement("div");
+    const props = createOverviewProps({
+      sessionsResult: {
+        ts: 0,
+        path: "",
+        count: 3,
+        defaults: { modelProvider: "openai", model: "gpt-5", contextTokens: null },
+        sessions: [
+          {
+            key: "discord:123:456",
+            kind: "direct",
+            label: "   ",
+            displayName: "Ops Room",
+            model: "gpt-5",
+            updatedAt: null,
+          },
+          {
+            key: "telegram:123:456",
+            kind: "direct",
+            label: "telegram:123:456",
+            model: "gpt-5",
+            updatedAt: null,
+          },
+          {
+            key: "agent:main:main",
+            kind: "direct",
+            label: "Main Project",
+            displayName: "agent:main:main",
+            model: "gpt-5",
+            updatedAt: null,
+          },
+        ],
+      },
+    });
+
+    render(renderOverview(props), container);
+    await Promise.resolve();
+
+    const recentNames = [...container.querySelectorAll(".ov-recent__key")].map(
+      (node) => node.textContent?.trim() ?? "",
+    );
+    expect(recentNames).toEqual(["Ops Room", "Telegram Session", "Main Project"]);
+    expect(recentNames).not.toContain("telegram:123:456");
+  });
 });
